CP-28114: Don't have duplicate uuids over a migrate
Instead, we create a domain on the destination with a uuid that is
related to the original VM uuid:

 xxxxxxxx-xxxx-xxxx-xxxx-000000000001

where the xs are from the original uuid. Once the domain has been
suspended on the source, we then change the uuid on the source to

 xxxxxxxx-xxxx-xxxx-xxxx-000000000000

at this point there is no domain with the VM's uuid anywhere. Then
the destination domain is renamed to the original uuid before it
is unpaused.

The commit also introduces aliasing to the thread queues, so we
can associate the work of one thread with multiple uuids. This is
to ensure that only one thread is dealing with the VM before and
after the uuid is changed and we don't race with the events thread.

Signed-off-by: Jon Ludlam <jonathan.ludlam@citrix.com>